docs(swagger): updated the unsign course route documentation
also refactored the implementation to fix the returns

+++ src/modules/users/useCases/unsignCourse/UnsignCourseUseCase.ts
@@ -1,6 +1,7 @@
 import { Course } from '@prisma/client';
 
 import { prisma } from '../../../../database';
+import { AppError } from '../../../../shared/errors/AppError';
 
 type IRequest = {
   userId: string;
@@ -9,6 +10,11 @@ type IRequest = {
 
 export class UnsignCourseUseCase {
   async execute({ userId, trailId }: IRequest): Promise<Course> {
+    const courseExists = await prisma.course.findFirst({ where: { trailId } });
+
+    if (!courseExists) {
+      throw new AppError('Course not found', 404);
+    }
     const course = await prisma.course.delete({
       where: {
         userId_trailId: {
